+++ include/nil/blueprint/components/hashes/poseidon/plonk/poseidon.hpp
@@ -188,7 +188,7 @@ namespace nil {
                 static_assert(state_size == 3);
 
                 for (std::size_t i = row; i < row + component.rows_amount - 1; i++) {
-                    for (std::uint32_t j = 0; j < state_size; j++) {
+                    for (std::size_t j = 0; j < state_size; j++) {
                         next_state[j] = state[0].pow(component_type::sbox_alpha) * component_type::mds[j][0] +
                                         state[1].pow(component_type::sbox_alpha) * component_type::mds[j][1] +
                                         state[2].pow(component_type::sbox_alpha) * component_type::mds[j][2] +
